LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-color-input.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;

function runTest() {
    const bounds = colorInputElement.getBoundingClientRect();
    range = document.caretRangeFromPoint(bounds.left + bounds.width / 2, bounds.top + bounds.height / 2);
    shouldBe("range.commonAncestorContainer", "colorInputElement");
    finishJSTest();
}
</script>
</head>
<body onload="runTest()">
    <input type="color" id="colorInputElement" style="position:absolute; height: 100px; width: 100px;">
</body>
</html>
