LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-color-input-expected.txt

PASS range.commonAncestorContainer is colorInputElement
PASS successfullyParsed is true

TEST COMPLETE
